backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
black==24.10.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
deprecation==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.2
flake8==7.1.1
flower==2.0.1
frozenlist==1.4.1
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
mjml-python==1.3.4
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
postgrest==0.17.1
premailer==3.10.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.5
redis==5.1.1
requests==2.32.3
resend==2.4.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
starlette==0.40.0
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
threadpoolctl==3.5.0
tomli==2.0.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
websockets==13.1
yarl==1.15.4
